bpo-36593: Fix isinstance check for Mock objects with spec executed under tracing (GH-12790)
In Python having a trace function in effect while mock is imported causes isinstance to be wrong for MagicMocks. This is due to the usage of super() in some class methods, as this sets the __class__ attribute. To avoid this, as a workaround, alias the usage of super .

Backports: 830b43d03cc47a27a22a50d777f23c8e60820867
Signed-off-by: Chris Withers <chris@withers.org>

A lot of munging needed:
- the fix was already present as it was needed to get a prior patch to work on Py2.
- the test needed adjustment due to the module location differences.